==> 01. Text Preprocessing.ipynb <==
# Amazon Fine Food Reviews Analysis

Data source : https://www.kaggle.com/snap/amazon-fine-food-reviews

The Amazon Fine Food Reviews dataset consists of reviews of fine foods from Amazon.

Number of reviews: 568,454
Number of users: 256,059
Number of Products: 74,258
Timespan: Oct 1999 - Oct 2012
Number of Attributes/Columns in data: 10

Attribute Information:
1. Id
2. ProductId - unique identifier for the product
3. UserId - unique identifier for the user
4. ProfileName
5. HelpfulnessNumerator - number of users who found the review helpful
6. HelpfulnessDenominator - number of users who indicated whether they found the review helpful or not
7. Score - rating between 1 and 5
8. Time - timestamp for the review
9. Summary - brief summary of the review
10. Text - text of the review

#### Objective

Given a review, determine whether the review is positive (Rating of 4 or 5) or negative (rating of 1 or 2).

[Q] How to determine if a review is positive or negative?

[Ans] We could use the Score/Rating. A rating of 4 or 5 could be considered a positive review. A review of 1 or 2 
could be considered negative. A review of 3 is neutral and therefore ignored. Note that this is an approximate and proxy way of determining the polarity (positivity/negativity) of a review.

## Loading the dataset

The dataset is available in two forms
1. .csv file
2. SQLite database

In order to load the data, We have used the SQLITE dataset as it is easier to query the data and visualise the data effectively.

Here as we only want to get the global sentiment of the recommendations (positive or negative), we will purposefully ignore all Scores equal to 3.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

import sqlite3

# Using the SQLite Table to read data.
con = sqlite3.connect('database.sqlite')

# Filtering only positive and negative reviews i.e. not taking
# into consideration those reviews whose score = 3
filtered_data = pd.read_sql_query('select * from reviews where score != 3 limit 100000', con)

print('Number of datapoints in our data {}'.format(filtered_data.shape))
filtered_data.head()

# Give reviews with a score > 3 a positive rating,
# and reviews with a score < 3 a negative rating
def partition(x):
    if x < 3:
        return 0 # negative
    return 1 # positive

actualScore = filtered_data['Score']
positiveNegative = actualScore.map(partition)
filtered_data['Score'] = positiveNegative
filtered_data.head()

## Exploratory Data Analysis

### Data Cleaning : Deduplication

It is observed (as shown in the table below) that the reviews data had many duplicate entries. Hence it was necessary to remove duplicates in order to get unbiased results for the analysis of the data.

Following is an example:

filtered_data[filtered_data['UserId'] == 'AR5J8UI46CURR']

- As can be seen above the same user has multiple reviews with the same values for HelpfulnessNumerator, HelpfulnessDenominator, Score, Time, Summary and Text and on doing analysis it was found that 

  ProductId=B000HDOPZG was Loacker Quadratini Vanilla Wafer Cookies, 8.82-Ounce Packages (Pack of 8)

  ProductId=B000HDL1RQ was Loacker Quadratini Lemon Wafer Cookies, 8.82-Ounce Packages (Pack of 8) and so on
  

- It was inferred after analysis that reviews with same parameters other than ProductId belonged to the same product just having different flavour or quantity. Hence in order to reduce redundency it was decided to eliminate the rows having same parameters.


- The method used to eliminate was that we first sort the data according to ProductId and then just keep the first similar product review and delete the others. For eg. in the above just the review for ProductId=B000HDL1RQ remains. This method ensures that there is only one representative for each product and deduplication without sorting would lead to possibility of different representatives still existing for the same product.

# Sorting data according to ProductId in ascending order
sorted_data = filtered_data.sort_values('ProductId', axis=0, ascending=True, inplace=False, kind='quicksort', na_position='last')

# Deduplication of entries
final = filtered_data.drop_duplicates(subset={"UserId", "ProfileName", "Time", "Text"}, keep='first', inplace=False)
final.shape

# Checking to see how much % of data still remains
(final.size / filtered_data.size) * 100

<b>Observation : </b>It was also seen that in two rows given below the value of HelpfulnessNumerator is greater than HelpfulnessDenominator which is not practically possible hence these two rows too are removed from calculations.

final[final['HelpfulnessNumerator'] > final['HelpfulnessDenominator']]

final = final[final['HelpfulnessNumerator'] <= final['HelpfulnessDenominator']]

# Before starting the next phase of preprocessing lets see the number of entries left
print(final.shape)

# How many positive and negative reviews are present in our dataset ?
final['Score'].value_counts()

## Text Preprocessing: Stemming, stop-word removal and Lemmatization

Now that we have finished deduplication our data requires some preprocessing before we go on further with analysis
and making the prediction model.

Hence in Preprocessing phase we do following in the order below:-

1. Begin by removing html tags
2. Remove any punctuations or limited set of special characters like , or . or # etc.
3. Check if the word is made up of english letters and is not alpha-numeric.
4. Check to see if the length of the word is greater than 2 (as it was researched that there is no adjective in 2-letters)
5. Convert the word to lowercase.
6. Remove Stopwords.
7. Finally Snowball stemming the word (it was observed to be better than Porter Stemming)

After which we collect the words used to describe positive and negative reviews.

# finding sentences containing HTML tags
import re
i = 0
for sent in final['Text'].values:
    if len(re.findall('<.*?>', sent)) > 0:
        print(i)
        print(sent)
        break
    i += 1

# defining some helper functions for text preprocessing
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

stop = set(stopwords.words('english')) # set of stopwords
sno = SnowballStemmer('english') # initialising the snowball stemmer

def cleanhtml(sentence): # function to clean html tags from sentence
    tag = re.compile('<.*?>')
    clean_sentence = re.sub(tag, ' ', sentence)
    return clean_sentence

def cleanpunc(sentence): # function to clean the sentence from any punctuation or special characters
    clean_sentence = re.sub(r'[?|!|\'|"|#]', r'', sentence)
    clean_sentence = re.sub(r'[.|,|)|(|\|/]', r' ', clean_sentence)
    return clean_sentence

print(stop)
print('***************************************************')
print(sno.stem('tasty'))

# Code for implementing step-by-step the checks mentioned in the pre-processing phase
# The code takes a while to run as it needs to run on 100k sentences.
import os
from tqdm import tqdm
if not os.path.isfile('final.sqlite'):
    i = 0
    cleaned_reviews = []
    all_positive_words = [] # store words from positive reviews here
    all_negative_words = [] # store words from negative reviews here
    for sent in tqdm(final['Text'].values):
        cleaned_words = [] # storing each cleaned word from sent
        sent = cleanhtml(sent) # remove html tags
        for w in sent.split():
            for cleaned_word in cleanpunc(w).split():
                if cleaned_word.isalpha() and len(cleaned_word)>2:
                    if cleaned_word.lower() not in stop: # stopword removal
                        s = (sno.stem(cleaned_word.lower())).encode('utf8') # stemming
                        cleaned_words.append(s)
                        if final['Score'].values[i] == 1:
                            all_positive_words.append(s) # list of all words used to describe positive reviews
                        if final['Score'].values[i] == 0:
                            all_negative_words.append(s) # list of all words used to describe negative reviews
                    else:
                        continue
                else:
                    continue
        str1 = b" ".join(cleaned_words) # final string of cleaned words
        cleaned_reviews.append(str1)
        i += 1

import pickle
if not os.path.isfile('final.sqlite'):
    ############-------- storing the preprocessed data into .sqlite file --------#############
    final['CleanedText'] = cleaned_reviews # adding a column of CleanedText which displays the data after pre-processing of the review
    final['CleanedText'] = final['CleanedText'].str.decode('utf-8')
    # store final table into a SQLite table for future.
    conn = sqlite3.connect('final.sqlite')
    c = conn.cursor()
    conn.text_factory = str
    final.to_sql('Reviews', conn, if_exists='replace')
    conn.close()
    
    with open('positive_words.pkl', 'wb') as f:
        pickle.dump(all_positive_words, f)
    with open('negative_words.pkl', 'wb') as f:
        pickle.dump(all_negative_words, f)
