==> 02. Bag of Words.ipynb <==
import os
import sqlite3
import pandas as pd
import nltk
import pickle
from sklearn.feature_extraction.text import CountVectorizer  # BOW implementation
import warnings
warnings.filterwarnings('ignore')

# Loading Dataset

if os.path.isfile('final.sqlite'):
    conn = sqlite3.connect('final.sqlite')
    final = pd.read_sql_query('select * from reviews', conn)
    conn.close()
else:
    print('Please run Text Preprocessing code file')

# Bag of Words

count_vect = CountVectorizer()
final_counts = count_vect.fit_transform(final['CleanedText'].values)
print("The type of text BOW vectorizer : ", type(final_counts))
print("The shape of text BOW vectorizer : ", final_counts.get_shape())
print("Number of unique words : ", final_counts.get_shape()[1])

## Bi-Grams and n-Grams

**Motivation**

Now that we have our list of words describing positive and negative reviews lets analyse them.

We begin analysis by getting the frequency distribution of the words as shown below

with open('positive_words.pkl', 'rb') as f:
    all_positive_words = pickle.load(f)
with open('negative_words.pkl', 'rb') as f:
    all_negative_words = pickle.load(f)

print(all_positive_words)
freq_dist_positive = nltk.FreqDist(all_positive_words) # list of words as arguments
freq_dist_negative = nltk.FreqDist(all_negative_words)
print("Most common positive words : ", freq_dist_positive.most_common(20))
print()
print("Most common negative words : ", freq_dist_negative.most_common(20))

<b>Observation:-</b> From the above it can be seen that the most common positive and the negative words overlap for eg. 'like' can be used as 'not like' etc.
So, it is a good idea to consider pairs of consequent words (bi-grams) or a sequence of n consecutive words (n-grams)

# bi-gram, tri-gram and n-gram
# removing stop words like "not" should be avoided before building n-grams
count_vect = CountVectorizer(ngram_range=(1,2)) # unigrams and bigrams (2,2) -> only bigrams
final_bigram_counts = count_vect.fit_transform(final['CleanedText'].values)
print("The type of text BOW vectorizer : ", type(final_bigram_counts))
print("The shape of text BOW vectorizer : ", final_bigram_counts.get_shape())
print("Number of unique words including both unigrams and bigrams : ", final_bigram_counts.get_shape()[1])
