==> 03. TF-IDF.ipynb <==
import os
import sqlite3
import pandas as pd
import numpy as np
import nltk
import pickle 
from sklearn.feature_extraction.text import TfidfVectorizer # Tfidf implementation
import warnings
warnings.filterwarnings('ignore')

# Loading Dataset

if os.path.isfile('final.sqlite'):
    conn = sqlite3.connect('final.sqlite')
    final = pd.read_sql_query('select * from reviews', conn)
    conn.close()
else:
    print('Please run Text Preprocessing code file')

# TF-IDF

tf_idf_vect = TfidfVectorizer(ngram_range=(1, 2))
final_tf_idf = tf_idf_vect.fit_transform(final['CleanedText'].values)
# final_tf_idf is a sparse matrix with row=sentence, col=word and cell_val=tfidf
print("The type of text Tfidf vectorizer : ", type(final_tf_idf))
print("The shape of text Tfidf vectorizer : ", final_tf_idf.get_shape())
print("Number of unique words including both unigrams and bigrams: ", final_tf_idf.get_shape()[1])

features = tf_idf_vect.get_feature_names()
print("some sample features(unique words in the corpus)\n", features[100:110])

# printing top n tfidf features
def top_tfidf_feats(row, features, top_n=25):
    ''' Get top n tfidf values in row and return them with their corresponding feature names.'''
    topn_ids = np.argsort(row)[::-1][:top_n]
    top_feats = [(features[i], row[i]) for i in topn_ids]
    df = pd.DataFrame(top_feats)
    df.columns = ['feature', 'tfidf']
    return df

toptfidf = top_tfidf_feats(final_tf_idf[1,:].toarray()[0], features, 25)

toptfidf
