==> 04. Word2Vec.ipynb <==
import sqlite3
import os
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
from gensim.models import KeyedVectors
from sklearn.feature_extraction.text import TfidfVectorizer
from tqdm import tqdm

# Loading Dataset

if os.path.isfile('final.sqlite'):
    conn = sqlite3.connect('final.sqlite')
    final = pd.read_sql_query('select * from reviews', conn)
    conn.close()
else:
    print('Please run Text Preprocessing code file')

# Word2Vec

# Method-1 -> Using Google News Word2Vectors

# In this project we are using a pretrained model by Google
# its 3.3GB file, once you load this into your memory it occupies
# ~9GB, so please do this step only if you have > 12GB of RAM.

# It contains all our courpus words as keys and  model[word] as values.

# To use this code-snippet, download "GoogleNews-vectors-negative300.bin" 
# from https://drive.google.com/file/d/0B7XkCwpI5KDYNlNUTTlSS21pQmM/edit
# it's 1.9GB in size.

if os.path.isfile('GoogleNews-vectors-negative300.bin'):
    model = KeyedVectors.load_word2vec_format('GoogleNews-vectors-negative300.bin', binary=True)

print("the vector representation of word 'computer' : \n", model['computer'])
print("the similarity between the words 'woman' and 'man' : \n", model.similarity('woman', 'man')) # partially similar
print("the most similar words to 'woman'", model.most_similar('woman'))

# Method-2 -> Train your own Word2Vec model using your own text corpus
list_of_sent = []
for sent in final['CleanedText'].values:
    list_of_sent.append(sent.split())

print(final['CleanedText'][0])
print("********************************************************")
print(list_of_sent[0])

# min_count = 2 considers only words that occured atleast 2 times
w2v_model = Word2Vec(list_of_sent, min_count=2, vector_size=50, workers=4)

w2v_words = list(w2v_model.wv.index_to_key)
print("number of words that occured minimum 2 times : ", len(w2v_words))
print("sample words : ", w2v_words[:50])

w2v_model.wv.most_similar('tasti')

w2v_model.wv.most_similar('like')

# Avg Word2Vec

# Average Word2Vec
# compute average word2vec for each review
sent_vectors = [] # the avg-w2v for each sentence/review is stored in this list
for sent in tqdm(list_of_sent): # for each review/sentence
    sent_vec = np.zeros(50) # as word vectors are of length 50
    cnt_words = 0 # num of words with a valid vector in the sentence/review
    for word in sent: # for each word in rview/sentence
        if word in w2v_words:
            vec = w2v_model.wv[word]
            sent_vec += vec
            cnt_words += 1
    if cnt_words != 0:
        sent_vec /= cnt_words # taking average
    sent_vectors.append(sent_vec)
print(len(sent_vectors))
print(len(sent_vectors[0]))

# TFIDF-Word2Vec

model = TfidfVectorizer()
tf_idf_matrix = model.fit_transform(final['CleanedText'].values)
# we are creating a dictionary with word as key, and its idf as value
dictionary = dict(zip(model.get_feature_names(), list(model.idf_)))

# TF-IDF weighted Word2Vec
tfidf_feat = model.get_feature_names() # tfidf words/col-names
# tf_idf_matrix is a sparse matrix with row=sentence, col=word and cell_val=tfidf

tfidf_sent_vectors = [] # the tfidf-w2v for each sentence/review is stored in this list
row=0
for sent in tqdm(list_of_sent): # for each review/sentence
    sent_vec = np.zeros(50) # as word vectors are of 50 length
    weight_sum = 0 # num of words with a valid vector in the sentence/review
    for word in sent: # for each word in a review/sentence
        if word in w2v_words:
            vec = w2v_model.wv[word]
            # tf_idf = tf_idf_matrix[row, tfidf_feat.index(word)]
            # to reduce the computations we are using below method to compute tf_idf
            # sent.count(word) = tf values of word in curr review
            # dictionary[word] = idf value of word in whole corpus
            tf_idf = (sent.count(word)/len(sent)) * dictionary[word]
            sent_vec += (vec * tf_idf)
            weight_sum += tf_idf
    if weight_sum != 0:
        sent_vec /= weight_sum
    tfidf_sent_vectors.append(sent_vec)
    row += 1
